+++ python/python/glide/async_commands/server_modules/ft_options/ft_constants.py
@@ -9,6 +9,9 @@ class CommandNames:
     FT_CREATE = "FT.CREATE"
     FT_DROPINDEX = "FT.DROPINDEX"
     FT_SEARCH = "FT.SEARCH"
+    FT_ALIASADD = "FT.ALIASADD"
+    FT_ALIASDEL = "FT.ALIASDEL"
+    FT_ALIASUPDATE = "FT.ALIASUPDATE"
 
 
 class FtCreateKeywords: